# use pubsub instead of filenotify to follow json logs