inotify event is trigged immediately there's data written to disk.
But at the time that the inotify event is received, the json line might
not fully saved to disk. If the json decoder tries to decode in such
case, an io.UnexpectedEOF will be trigged.
We used to retry for several times to mitigate the io.UnexpectedEOF error.
But there are still flaky tests caused by the partial log entries.

The daemon knows exactly when there are new log entries emitted. We can
use the pubsub package to notify all the log readers instead of inotify.

BTW: I can't run the test locally because the testing infrastructure seems to be broken on my laptop.

```
---> Making bundle: test-integration-cli (in bundles/1.10.0-dev/test-integration-cli)
---> Making bundle: .integration-daemon-start (in bundles/1.10.0-dev/test-integration-cli)
+++ /etc/init.d/apparmor start
 * Starting AppArmor profiles
Skipping profile in /etc/apparmor.d/disable: usr.sbin.rsyslogd
   ...done.
INFO: Waiting for daemon to start...
+++ exec docker daemon --debug --host unix:///go/src/github.com/docker/docker/bundles/1.10.0-dev/test-integration-cli/docker.sock --storage-driver vfs --pidfile bundles/1.10.0-dev/test-integration-cli/docker.pid --userland-proxy=true
.
---> Making bundle: .integration-daemon-setup (in bundles/1.10.0-dev/test-integration-cli)
---> Making bundle: .detect-daemon-osarch (in bundles/1.10.0-dev/test-integration-cli)
---> Making bundle: .ensure-emptyfs (in bundles/1.10.0-dev/test-integration-cli)
++++ tar -cC bundles/1.10.0-dev/test-integration-cli/emptyfs .
++++ docker load
---> Making bundle: .ensure-frozen-images (in bundles/1.10.0-dev/test-integration-cli)
++++ tar -cC /docker-frozen-images .
++++ docker load
Error response from daemon: ApplyLayer exit status 1 stdout:  stderr: unexpected EOF
+++ bundle .integration-daemon-stop
+++ local bundle=.integration-daemon-stop
/go/src/github.com/docker/docker/hack/make.sh: line 257: local: can only be used in a function
```

Signed-off-by: Shijiang Wei <mountkin@gmail.com>