@cpuguy83 2016-01-22
I like the general simplification, but the issue is now for every write we have to notify as well.
I think the flakey tests are generally from a race between stopping the log writer (on container stop) and getting all that out of the buffer in time.