@mountkin 2016-01-22
@cpuguy83 I think the publish is much cheaper than inotify because if no one is running 'docker logs -f' the 'Publish' function does nothing more than locking and unlocking a mutex. And the 'Publish' is only called when a line is fully written to the disk (while inotify might be triggered many times when writing a line, see my debug log in another PR https://github.com/docker/docker/pull/17606).
If someone is running, the subscriber is also very lightweight and it never blocks (merely a atomic.AddUint64).