@cpuguy83 2016-01-28
The extra call for every write even when there may not be a reader is a little iffy.
The thing with inotify is, it's only consuming resources when there is a reader.

```
BenchmarkPubSub-8  	10000000	       175 ns/op	     260 B/op	       1 allocs/op
BenchmarkNoPubSub-8	30000000	        82.1 ns/op	     163 B/op	       0 allocs/op
```

This is comparing just writing with the pubsub call and writing without.