@mountkin 2016-01-28
@cpuguy83 the current version of 'Publish' function can be optimized a little. If no subscribers the waitgroup allocation can be omitted.