@mountkin 2016-01-29
@cpuguy83 Here is the result of my benchmark:
`go test -v -benchtime 10s -test.bench BenchmarkJSONFileLogger -test.benchmem -test.run=BenchmarkJSONFileLogger` 

filenotify:
```BenchmarkJSONFileLogger-8	  200000	     60936 ns/op	  59.57 MB/s	     960 B/op	      30 allocs/op```

before https://github.com/mountkin/docker/commit/319b7827eb864b56e2b28baf12b4955aa1fff0e8:
```BenchmarkJSONFileLogger-8	  200000	     68881 ns/op	  52.70 MB/s	    1440 B/op	      60 allocs/op```

after https://github.com/mountkin/docker/commit/319b7827eb864b56e2b28baf12b4955aa1fff0e8:
```BenchmarkJSONFileLogger-8	  200000	     64656 ns/op	  56.14 MB/s	     960 B/op	      30 allocs/op```

Is the overhead after `pubsub` optimization acceptable?