@cpuguy83 2016-02-06
Thanks @mountkin!

Yep, I'm good with those results, but I think we need some simplifications in the code.