@mountkin 2016-02-13
@thaJeztah sorry for the late response. I'm on holiday these days. I'll look into the failed tests tomorrow.