@mountkin 2016-02-15
Updated. 
@cpuguy83 PTAL this PR and https://github.com/docker/docker/issues/20327