@mountkin 2016-02-15
BTW, it seems that the `filenotify` package is no longer needed when this PR is merged. Can I remove the package in another commit?