@cpuguy83 2016-02-16
@mountkin Let's save it for another PR I think, b/c it's going to make this PR difficult to look at, even if in different commits.