@thaJeztah 2016-02-17
Based on https://github.com/docker/docker/issues/20327#issuecomment-185019751, this is a regression in 1.10.x, so I'm adding this to the 1.10.2 milestone