@anusha-ragunathan 2016-02-17
ARM tests are failing in TestLogsFollowSlowStdoutConsumer due to this change.
https://jenkins.dockerproject.org/job/Docker%20Master%20%28arm%29/694/console