@tiborvass 2016-02-19
I'm removing this from 1.10.2 as this changes a lot and we'd rather not add more risk than necessary to 1.10.2