@cpuguy83 2016-02-24
Yeah, this change is definitely slower :(
The existing benchmark is only really testing the marshalling and not really the full write->read flow

```
benchmark                 old ns/op     new ns/op     delta
BenchmarkJSONLogger-8     2162          2938          +35.89%

benchmark                 old allocs     new allocs     delta
BenchmarkJSONLogger-8     7              9              +28.57%

benchmark                 old bytes     new bytes     delta
BenchmarkJSONLogger-8     245           336           +37.14%
```