@cpuguy83 2016-02-24
And that's with the file poller too, not inotify.